BUG: Tiles are out of alignment or outside the bounds of the grid, sometimes overlapping.
Steps: If you swipe fast enough in random directions, you can usually see this bug. It was easier to produce when I slowed the animation of the moving tile down.

Issue: The issue is that the mainscene was processing touch events from the swipes before the animation cycle was complete. This would cause the tiles to update their positions mid move, thus causing the tiles to be out of alignment.

Solution: The solution was to not process swipe move events until the tiles were done moving from the previous round.

Added a callback whenever a tile is called. This is part of a animation sequence, so when this callback is called I know that the animation has completed. There is an atomic int counter that gets decremented because the animation has completed. Basically, there is a member variable that keeps track of completed animations. Everytime a tile move happens, the variable is incremented. Everytime an animation finishes, it decrements. Whenever touch input is initiated, there is a check to see if this counter is 0. If it is, then proceed to process touch input. Otherwise, toss out the input because the grid is not ready to process it.