BUG:
This fix addresses the bug with the gameover/win screens when the animations for create and explosions would freeze mid animation leaving an unsightly partial frame update in view. The general solution is to have all the animations complete before presenting the gameover screen. In creating the logic, I came across an existing memory leak bug where there would be orphaned action objects that would never execute because the target object has been removed before the action could initiate. Part of the solution was to create a container for all Action objects so that we would know when the actions were done executing. Once the container has been flushed (size 0) then we know we have no more actions to initiate in the queue, and so it is safe to present the gameover dialog. (As a side note, lambda expressions are very powerful and with this fix I actually use them for the first time. Will be using them in the future). Part of the problem was that the action objects would be automatically reclaimed by the memory management built into Cocos2Dx. I had to explicity retain() and release() these objects to prevent this from happening. Added tag identifiers to the actions to more easily identify which action I am dealing with. Also disabled input processing when the dialog pops up. In order to address the orphaned actions, I created a callback mechanism that would fire BEFORE the target object would be removed, so that I had an opportunity to cleanup, a.k.a remove the creation animation action from the member container. Now the container will flush correctly without orphaned actions hanging around. Added some init calls in order to get access to the update() for every frame since we need to check the action container every frame.