Bug: Icons were unsightly.
Updated all the icons of the 6 different resolutions for Android so that they are consistent with iOS look and feel. Basically, found and applied an iOS7 PNG 1024x1024 graphic that I used as a clipping mask on the 1024x1024 version of the icon. I then exported that graphic with a bicupic sharpen to retain the details without aliasing. This export is recommended when dealing with size reduction. The icons look like the iOS icons on the android device.